+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Python dataset queries can now filter annotation evaluation results for false positives and false negatives.
 - Show numeric metadata values in distribution panel.
 - Add a selector for the numeric metadata histogram bin count.
+- Introduce button to see expanded distribution for numeric metadata values.
 
 
 ### Changed